requirements.txt
fastapi>=0.115.0
uvicorn>=0.29.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.2.0
python-dotenv>=1.0.1
pandas>=2.2.3
scikit-learn>=1.5.2
pydantic[email]>=2.9.2
requests==2.32.3
google-generativeai==0.7.2




